+++ changelog.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.0] - 2025-09-10
+
+### Changed
+
+- Statstidende certificate moved to Keyvault.
+- Boliglån module uses uiautomation instead of pywinauto.
+- Moved email text from argument to config file.
+
+### Fixed
+
+- Opus emails are now deleted after the process is done.
 
 ## [1.0.2] - 2025-08-18
 
@@ -23,6 +33,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[Unreleased]: https://github.com/itk-dev-rpa/Gennemsoegning-af-Statstidende/compare/1.0.1...HEAD
+[1.1.0]: https://github.com/itk-dev-rpa/Gennemsoegning-af-Statstidende/releases/tag/1.1.0
 [1.0.1]: https://github.com/itk-dev-rpa/Gennemsoegning-af-Statstidende/releases/tag/1.0.1
 [1.0.0]: https://github.com/itk-dev-rpa/Gennemsoegning-af-Statstidende/releases/tag/1.0.0